Still doesn't quite work.. but closer
Shows the dead screen on loading it in processing IDE.
It was really about variable declaration.
Also... this program could really use a whole bunch of arrays! Get your other group members to looks at it and see if they are willing to clean it up a little. There is a lot to learn, and a lot to improve. Only because there is a lot of work put into it!
Keep up the good work!